Ionic App Activity TXT
The activity is an ionic app that has:
Tab 1 being "Home" that contains info about myself, along with my picture, age, school, and bio.
Tab 2 being "Blogs" that allows for blogs to be created and posted, as well as display all of the blogs and have a form to input new blogs.
Tab 3 being "Friends" that display a list of most of my friends and what they do. I pushed away the idea of posting
contact info as this assignment is public and I do not want their information to be out there without permission.